First version committed
This is a very simple first pass, with known issues.  It doesn't directly
query servers, only looks at a test file in /tmp/ with example data, and
assumes a set server name.  It also isn't doing proper matching of package
versions, since I'm currently testing on my laptop and can't use the RPM
gem yet.  However, it does read in packages and compares to the centos
advisories, then prints out a server report.

It's a start.