Add tests for the report builder class
The data being tossed at it right now is too simple and should be expanded
out in the long term, but for right now we have some basic testing of the
report data.  This checks on the data returned by the report functions,
rather than looking at the formatting that's then applied by specific reports.