Rewrote importing advisories for RHEL
The CVRF source for RHEL wasn't the best for our use.  It only reported
on the source packages affected by an advisory, which caused problems
when a source package built multiple packages (ex: bind building bind and
bind-lib and others).

Moved to use the OVAL sources instead, which do contain package
information directly.